refactor: 경로조회 인수테스트 적절한 위치로 이동
