+++ src/test/java/nextstep/subway/path/PathAcceptanceTest.java
@@ -1,9 +1,147 @@
 package nextstep.subway.path;
 
+import static nextstep.subway.line.acceptance.LineAcceptanceTest.지하철_노선_등록되어_있음;
+import static nextstep.subway.line.acceptance.LineSectionAcceptanceTest.지하철_노선에_지하철역_등록_요청;
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.junit.jupiter.api.Assertions.assertAll;
+
+import io.restassured.RestAssured;
+import io.restassured.response.ExtractableResponse;
+import io.restassured.response.Response;
+import java.util.Arrays;
+import java.util.List;
+import java.util.stream.Collectors;
 import nextstep.subway.AcceptanceTest;
+import nextstep.subway.line.dto.LineRequest;
+import nextstep.subway.line.dto.LineResponse;
+import nextstep.subway.line.dto.PathResponse;
+import nextstep.subway.station.StationAcceptanceTest;
+import nextstep.subway.station.dto.StationResponse;
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.MediaType;
 
 
 @DisplayName("지하철 경로 조회")
 public class PathAcceptanceTest extends AcceptanceTest {
+    private LineResponse 신분당선;
+    private LineResponse 이호선;
+    private LineResponse 삼호선;
+    private StationResponse 강남역;
+    private StationResponse 양재역;
+    private StationResponse 교대역;
+    private StationResponse 남부터미널역;
+
+    /**
+     * 교대역    --- *2호선* ---   강남역
+     * <p>
+     * |                        |
+     * <p>
+     * *3호선*                   *신분당선*
+     * <p>
+     * |                        |
+     * <p>
+     * 남부터미널역  --- *3호선* ---   양재
+     */
+    @BeforeEach
+    public void setUp() {
+        super.setUp();
+
+        강남역 = StationAcceptanceTest.지하철역_등록되어_있음("강남역").as(StationResponse.class);
+        양재역 = StationAcceptanceTest.지하철역_등록되어_있음("양재역").as(StationResponse.class);
+        교대역 = StationAcceptanceTest.지하철역_등록되어_있음("교대역").as(StationResponse.class);
+        남부터미널역 = StationAcceptanceTest.지하철역_등록되어_있음("남부터미널역").as(StationResponse.class);
+
+        신분당선 = 지하철_노선_등록되어_있음(new LineRequest("신분당선", "bg-red-600", 강남역.getId(), 양재역.getId(), 10)).as(
+                LineResponse.class);
+        이호선 = 지하철_노선_등록되어_있음(new LineRequest("이호선", "bg-red-600", 교대역.getId(), 강남역.getId(), 10)).as(
+                LineResponse.class);
+        삼호선 = 지하철_노선_등록되어_있음(new LineRequest("삼호선", "bg-red-600", 교대역.getId(), 양재역.getId(), 5)).as(
+                LineResponse.class);
+
+        지하철_노선에_지하철역_등록_요청(삼호선, 교대역, 남부터미널역, 3);
+    }
+
+    /**
+     * when 출발역, 도착역을 지정하여 최단 경로를 조회하면
+     * <p>
+     * then 최단경로가 조회됨
+     */
+    @DisplayName("지하철 최단 경로를 조회한다")
+    @Test
+    void findShortestPath() {
+        // when
+        ExtractableResponse<Response> response = 지하철_경로_조회_요청(교대역, 양재역);
+
+        // then
+        최단_거리_확인됨(response, Arrays.asList(교대역, 남부터미널역, 양재역), 5);
+    }
+
+    /**
+     * when 출발역과 도착역이 같게하여, 최단 경로를 조회하면
+     * <p>
+     * then 최단경로를 조회할 수 없다.
+     */
+
+    @DisplayName("출발역과 도착역이 같은 경우, 최단경로를 조회 할 수 없다")
+    @Test
+    void findSourceAndTargetSameShortestPath() {
+        // when
+        ExtractableResponse<Response> response = 지하철_경로_조회_요청(교대역, 교대역);
+
+        // then
+        최단_경로_조회_실패됨(response);
+    }
+
+    /**
+     * when 존재하지 않는 경로를 조회 할 경우
+     * <p>
+     * then 최단경로를 조회할 수 없다.
+     */
+
+    @DisplayName("존재하지 않는 경로를 조회 할 경우, 최단경로를 조회 할 수 없다")
+    @Test
+    void findNotExistedPath() {
+        // when
+        StationResponse 없는역 = StationAcceptanceTest.지하철역_등록되어_있음("없는역").as(StationResponse.class);
+        ExtractableResponse<Response> response = 지하철_경로_조회_요청(교대역, 없는역);
+
+        // then
+        최단_경로_조회_실패됨(response);
+    }
+
+    public ExtractableResponse<Response> 지하철_경로_조회_요청(StationResponse 출발역, StationResponse 도착역) {
+        return RestAssured
+                .given().log().all()
+                .accept(MediaType.APPLICATION_JSON_VALUE)
+                .when().get("/paths?source={sourceId}&target={targetId}", 출발역.getId(), 도착역.getId())
+                .then().log().all()
+                .extract();
+    }
+
+    public static void 최단_거리_확인됨(ExtractableResponse<Response> 지하철_경로_조회_결과, List<StationResponse> expectedStations,
+                                 int expectedDistance) {
+
+        PathResponse response = 지하철_경로_조회_결과.as(PathResponse.class);
+
+        List<Long> stationIds = response.getStations().stream()
+                .map(StationResponse::getId)
+                .collect(Collectors.toList());
+
+        List<Long> expectedStationIds = expectedStations.stream()
+                .map(it -> it.getId())
+                .collect(Collectors.toList());
+
+        assertAll(
+                () -> assertThat(response.getDistance()).isEqualTo(expectedDistance),
+                () -> assertThat(stationIds).containsExactlyElementsOf(expectedStationIds)
+        );
+
+    }
+
+    public static void 최단_경로_조회_실패됨(ExtractableResponse<Response> response) {
+        assertThat(response.statusCode()).isEqualTo(HttpStatus.INTERNAL_SERVER_ERROR.value());
+    }
 }
